feat: implement favorite APIs and extract logic into controller
