+++ package.json
@@ -25,10 +25,11 @@
     "ecpay_aio_nodejs": "^1.2.2",
     "express": "^5.1.0",
     "faker": "^6.6.6",
-    "jsonwebtoken": "^9.0.0",
     "minimist": "^1.2.8",
     "nanoid": "^3.3.4",
+    "jsonwebtoken": "^9.0.2",
     "minimist": "^1.2.8",
+    "nanoid": "^3.3.4",
     "nodemon": "^3.1.10",
     "pg": "^8.16.0",
     "zod": "^3.25.56"
